+++ geolocation/android/src/main/java/com/capacitorjs/plugins/geolocation/Geolocation.java
@@ -3,6 +3,7 @@ package com.capacitorjs.plugins.geolocation;
 import android.content.Context;
 import android.location.Location;
 import android.location.LocationManager;
+import android.os.SystemClock;
 import androidx.core.location.LocationManagerCompat;
 import com.google.android.gms.location.FusedLocationProviderClient;
 import com.google.android.gms.location.LocationCallback;
@@ -83,5 +84,25 @@ public class Geolocation {
             fusedLocationClient.removeLocationUpdates(locationCallback);
             locationCallback = null;
         }
+    }
+
+    @SuppressWarnings("MissingPermission")
+    public Location getLastLocation(int maximumAge) {
+        Location lastLoc = null;
+        LocationManager lm = (LocationManager) context.getSystemService(Context.LOCATION_SERVICE);
+        for (String provider : lm.getAllProviders()) {
+            Location tmpLoc = lm.getLastKnownLocation(provider);
+            if (tmpLoc != null) {
+                long locationAge = SystemClock.elapsedRealtimeNanos() - tmpLoc.getElapsedRealtimeNanos();
+                long maximumAgeNanoSec = maximumAge * 1000000L;
+                if (
+                    locationAge <= maximumAgeNanoSec &&
+                    (lastLoc == null || lastLoc.getElapsedRealtimeNanos() > tmpLoc.getElapsedRealtimeNanos())
+                ) {
+                    lastLoc = tmpLoc;
+                }
+            }
+        }
+        return lastLoc;
     }
 }